+++ src/core/converter/aerofly-flight/AeroflyFlightToGeoJsonConverter.ts
@@ -16,10 +16,12 @@ export class AeroflyFlightToGeoJsonConverter extends AeroflyFlightToStringConver
     convert(flightplan: AeroflyFlight): string {
         const geoJson = new FeatureCollection();
 
+        let id = 0;
+
         flightplan.navigation.waypoints.forEach((wp, index) => {
             geoJson.addFeature(
                 new Feature(this.getPointForWaypoint(wp), {
-                    id: index,
+                    id: id++,
                     sequence: index + 1,
                     title: wp.identifier,
                     type: wp.type,
@@ -28,8 +30,6 @@ export class AeroflyFlightToGeoJsonConverter extends AeroflyFlightToStringConver
                 }),
             );
         });
-
-        const index = flightplan.navigation.waypoints.length + 1;
 
         geoJson.addFeature(
             new Feature(
@@ -44,7 +44,7 @@ export class AeroflyFlightToGeoJsonConverter extends AeroflyFlightToStringConver
                     }),
                 ]),
                 {
-                    id: index + 1,
+                    id: id++,
                     title: this.getFlightplanTitle(flightplan),
                     type: "flightplan",
                     stroke: "#FF1493",
@@ -62,7 +62,7 @@ export class AeroflyFlightToGeoJsonConverter extends AeroflyFlightToStringConver
                 {
                     title: flightplan.aircraft.name,
                     livery: flightplan.aircraft.paintscheme,
-                    id: index + 2,
+                    id: id++,
                     type: "aircraft_position",
                     altitude_ft: flightplan.flightSetting.altitude_ft,
                     "marker-symbol": "airfield",